+++ app/assets/javascripts/discourse/app/deprecation-workflow.js
@@ -1,20 +1,6 @@
 import setupDeprecationWorkflow from "ember-cli-deprecation-workflow";
-
-const DEPRECATION_WORKFLOW = [
-  {
-    handler: "silence",
-    matchId: "ember-this-fallback.this-property-fallback",
-  },
-  { handler: "silence", matchId: "discourse.select-kit" },
-  { handler: "silence", matchId: "discourse.d-section" },
-  {
-    handler: "silence",
-    matchId: "discourse.decorate-widget.hamburger-widget-links",
-  },
-];
+import DEPRECATION_WORKFLOW from "discourse-common/deprecation-workflow";
 
 // We're using RAISE_ON_DEPRECATION in environment.js instead of
 // `throwOnUnhandled` here since it is easier to toggle.
 setupDeprecationWorkflow({ workflow: DEPRECATION_WORKFLOW });
-
-export default DEPRECATION_WORKFLOW;
